update UI Map page
