+++ frontend/src/App.css
@@ -310,3 +310,7 @@ transform: translate(-50%, 30%);
 .ant-tabs-tab {
   text-align: center;
 }
+
+.btn-curious-colors {
+  background-color: #7958f5
+}
