+++ frontend/src/pages/map/style.css
@@ -1,0 +1,56 @@
+.header-map {
+  padding: 20px;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  position: relative;
+}
+
+.header-map__navigate-left {
+  font-size: 25px;
+  position: absolute;
+  left: 10px
+}
+
+.header-map__title {
+  font-size: 17px;
+  font-weight: bold;
+}
+
+.header-map__navigate-right {
+  font-size: 25px;
+  position: absolute;
+  right: 10px
+}
+
+.footer-map {
+  width: 100%;
+  padding: 20px 30px;
+}
+
+.footer-map__btn-confirm {
+  width: 100%;
+  border-radius: 10px;
+  font-size: 16px;
+  text-align: center;
+  justify-content: center;
+}
+
+.ant-slider-track {
+  background-color: #7f57ff;
+}
+
+.ant-slider-track:focus {
+  background-color: #7f57ff;
+}
+
+.ant-slider-handle {
+  border: solid 2px #7f57ff;
+}
+
+.ant-slider-handle:focus {
+  border: solid 2px #7f57ff;
+}
+
+
+
